Organize classes into folders and remove debug printing methods
